bundle name changes
Hey guys, I'm following the guide using Webpack 5.52.0 and found inconsistencies (maybe because this version doesn't have the issues that are solved in the guide).

"If we run another build without making any changes, we'd expect that filename to stay the same. However, if we were to run it again, we may find that this is not the case": In my case, the filename actually stays the same.

"Running another build, we would expect only our main bundle's hash to change, however...": In my case this is actually what happens, only the main bundle's hash changes.

I'm not sure if the guide should be updated to reflect the behaviour of Webpack 5.